Fix delete records.
delete_record() threw errors and used to work by replacing all records.
This patch sends a proper DELETE request instead.

`record_type` has been kept as an optional parameter to avoid breaking
API compatibility, but leaving it unspecified does not delete all records
with a given name anymore (there is no DELETE endpoint for that).
Instead, it defaults to 'A'.