Add support for pagination.
The maximum pagination limit is 1000 for domains and 500 for records,
which are set as the defaults for the new function parameters.

For get_domains(), explicit parameters were added to make it clear that
pagination applies to this call, even though the user could already pass
arbitrary parameters. The same parameters were added to get_records(),
for which pagination was not possible (since no user-defined parameters
are accepted).